Fix lost performer input values
